[Binary Trees]-Diameter of a Binary Tree
For a given Binary of type integer, find and return the ‘Diameter’.

The diameter of a tree can be defined as the maximum distance between two leaf nodes.
Here, the distance is measured in terms of the total number of nodes present along the path of the two leaf nodes, including both the leaves.