[Binary Search Trees]-Find LCA
Given a binary tree and data of two nodes, find 'LCA' (Lowest Common Ancestor) of the given two nodes in the binary tree

LCA of two nodes A and B is the lowest or deepest node which has both A and B as its descendants.
It is defined that each node is a descendant to itself, so, if there are two nodes X and Y and X has a direct connection from Y, then Y is the lowest common ancestor.