[HashMaps]-Longest Consecutive Sequence
You are given an array of unique integers that contain
numbers in random order. You have to find the longest
possible sequence of consecutive numbers using the
numbers from given array.
You need to return the output array which contains
starting and ending element. If the length of the longest
possible sequence is one, then the output array must
contain only single element.