[Priority Queues]-Maximum Priority Queue Class
Implement the class for Max Priority Queue which includes
following functions -

1. getSize -
Return the size of priority queue i.e. number of elements
present in the priority queue.

2. isEmpty -
Check if priority queue is empty or not. Return true or false
accordingly.

3. insert -
Given an element, insert that element in the priority queue
at the correct position.

4. getMax -
Return the maximum element present in the priority queue
without deleting. Return -Infinity if priority queue is empty.

5. removeMax -
Delete and return the maximum element present in the
priority queue. Return -Infinity if priority queue is empty.