[Dynamic Programming]-Minimum Cost Path using DP
An integer matrix of size (M x N) has been given. Find out
the minimum cost to reach from the cell (0, 0) to (M - 1, N -
1).
From a cell (i, j), you can move in three directions:
1. ((i + 1), j) which is, "down"
2. (i, (i + 1)) which is, "to the right"
3. ((i+1), (j+1)) which is, "to the diagonal"
The cost of a path is defined as the sum of each cell's
values through which the route passes.