[Dynamic Programming]-Longest Consecutive Sequence Using DP
Given two strings, 'S' and 'T' with lengths 'M' and 'N', find
the length of the 'Longest Common Subsequence'.
For a string 'str' (per se) of length K, the subsequences are
the strings containing characters in the same relative order
as they are present in 'str,' but not necessarily contiguous.
Subsequences contain all the strings of length varying
from 0 to K.