[Dynamic Programming]-Minimum Number of Chocolates
Miss. Noor Rashid is a teacher. She wants to give some
chocolates to the students in her class. All the students sit
in a line, and each of them has a score according to
performance. Noor wants to give at least one chocolate to
each student. She distributes chocolates to them such that
If two students sit next to each other, then the one with the
higher score must get more chocolates. Miss. Noor wants
to save money, so she wants to minimize the total number
of chocolates.