[Graphs]-Get Path DFS
Given an undirected graph G(V, E) and two vertices v1 and
v2 (as integers), find and print the path from v1 to v2 (if
exists). Print nothing if there is no path between v1 and v2.
Find the path using DFS and print the first path that you
encountered.